BACKPORT: msm: kgsl: Fix preempt irq race with slumber
For no-gmu targets, we have a check for pending irq in the slumber
sequence. But, this is missing in gmu based targets. This leaves a
window open for prematurely moving to slumber state when kgsl hasn't
served all the pending interrupts. And that can leave some internal
kgsl states like preemption state incorrect. So, add a check for pending
irq and skip slumber if there is anything pending.

While we are at it, refactor the idle_check function to remove the deep
nesting to make it more readable.

Change-Id: I2cd53ba151dadbd5d50dc406a75ff1415ac0dad9
Signed-off-by: Akhil P Oommen <quic_akhilpo@quicinc.com>
[Tashar02: a6xx_irq_pending() -> adreno_irq_pending() on msm-5.4]
Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>
(cherry picked from commit 15890d27d05ebabeba384451c23166dfa71572de)